+++ src/containers/isHuman.js
@@ -1,0 +1,8 @@
+export default (player) => {
+  const attack = (coordinate) => {
+    // opponent.receiveAttack(coordinate);
+    player.getOpponentBoard().receiveAttack(coordinate);
+    // console.log(player.getOpponentBoard().board);
+  };
+  return { attack, ...player };
+};