+++ src/index.js
@@ -2,6 +2,6 @@ import './index.css';
 import Gameboard from './containers/gameboard';
 
 const foo = Gameboard();
-// foo.placeShip([5, 3], false);
 foo.placeShip([2, 8], true);
+foo.placeShip([5, 3], false);
 console.log(foo.board);