+++ webpack.config.js
@@ -39,4 +39,4 @@ module.exports = {
   optimization: {
     runtimeChunk: 'single',
   },
-}
+};